+++ spec/dummy/components/system-test-focused-view.tsx
@@ -20,8 +20,6 @@ export function SystemTestFocusedView({
       style={[{ flex: 1 }, style]}
       testID="systemTestingComponent"
       nativeID="systemTestingComponent"
-      accessibilityLabel="systemTestingComponent"
-      accessible
       collapsable={false}
       // @ts-ignore dataSet is supported at runtime for React Native Web
       dataSet={{ focussed: isFocussed ? 'true' : 'false' }}